Move PAX component and export it properly.
